Use other modules instead of re-implemented codes
