--- lib/is-string.js
@@ -1,15 +1,0 @@
-'use strict';
-
-function isString(value) {
-  if (typeof value === 'string') {
-    return true;
-  }
-
-  if (Object.prototype.toString.call(value) === '[object String]') {
-    return true;
-  }
-
-  return false;
-}
-
-module.exports = isString;
